requirements.txt
torch>=2.2
torchvision>=0.17
streamlit>=1.36
Pillow>=10.0
numpy>=1.26
tqdm>=4.66